Create song menu scene

+++ MusicRhythmGame/Assets/Scripts/MinionManager.cs
@@ -22,7 +22,9 @@ public class MinionManager : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        if (playerTransform.position.z - safeZone > (spawnZ - amnTilesOnScreen * tileLength)) {
+        GameObject soundManager = GameObject.Find("SoundManager");
+        bool isMusicEnding = soundManager.GetComponent<SoundManager>().IsMusicEnding();
+        if (!isMusicEnding && (playerTransform.position.z - safeZone > (spawnZ - amnTilesOnScreen * tileLength))) {
             SpawnMinion();
             DeleteMinion();
         }
